requirements.txt
#------------------------------------------------------------------------------
#--- !!!   IMPORTANT   !!!
#    This is a Collection of all Project Dependencies.
#    !!! DO NOT REMOVE !!!
#------------------------------------------------------------------------------
#--- Upgrade all
#    pip freeze --local | grep -v "^\-e" | cut -d = -f 1 | xargs -n1 pip install -U

#------------------------------------------------------------------------------
# --- "#"
git+ssh://git@github.com/asuvorov/ddaemon-core-python.git@dev

#------------------------------------------------------------------------------
# --- "A"
aioredis==2.0.1

# -----------------------------------------------------------------------------
# --- "B"

# -----------------------------------------------------------------------------
# --- "C"
channels==4.0.0
channels-redis==4.0.0

# -----------------------------------------------------------------------------
# --- "D"

# -----------------------------------------------------------------------------
# --- "django-*"
django-redis==5.2.0

#------------------------------------------------------------------------------
# --- "E"

#------------------------------------------------------------------------------
# --- "F"

#------------------------------------------------------------------------------
# --- "G"
gunicorn==20.1.0

#------------------------------------------------------------------------------
# --- "H"
httpx==0.24.0

#------------------------------------------------------------------------------
# --- "I"

#------------------------------------------------------------------------------
# --- "J"
jwt==1.3.1

#------------------------------------------------------------------------------
# --- "K"

#------------------------------------------------------------------------------
# --- "L"

#------------------------------------------------------------------------------
# --- "M"

#------------------------------------------------------------------------------
# --- "N"

#------------------------------------------------------------------------------
# --- "O"

#------------------------------------------------------------------------------
# --- "P"

# -----------------------------------------------------------------------------
# --- "python-*"

#------------------------------------------------------------------------------
# --- "Q"

#------------------------------------------------------------------------------
# --- "R"

# -----------------------------------------------------------------------------
# --- "S"

#------------------------------------------------------------------------------
# --- "T"

#------------------------------------------------------------------------------
# --- "U"
uvicorn[standard]

#------------------------------------------------------------------------------
# --- "V"

#------------------------------------------------------------------------------
# --- "W"
websocket-client==1.5.1

#------------------------------------------------------------------------------
# --- "X"

#------------------------------------------------------------------------------
# --- "Y"

#------------------------------------------------------------------------------
# --- "Z"
